Switched XRITLibraryTest to LibraryTest (Only console now)
