--- XRITLibraryTest/Program.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using Gtk;
-
-namespace XRITLibraryTest {
-    class MainClass {
-        public static void Main(string[] args) {
-            Application.Init();
-            MainWindow win = new MainWindow();
-            win.Show();
-            Application.Run();
-        }
-    }
-}
