--- XRITLibraryTest/gtk-gui/MainWindow.cs
@@ -1,37 +1,0 @@
-
-// This file has been generated by the GUI designer. Do not modify.
-
-public partial class MainWindow
-{
-	private global::Gtk.Fixed fixed1;
-	
-	private global::Gtk.FileChooserButton fileChooser;
-
-	protected virtual void Build ()
-	{
-		global::Stetic.Gui.Initialize (this);
-		// Widget MainWindow
-		this.Name = "MainWindow";
-		this.Title = global::Mono.Unix.Catalog.GetString ("MainWindow");
-		this.WindowPosition = ((global::Gtk.WindowPosition)(4));
-		// Container child MainWindow.Gtk.Container+ContainerChild
-		this.fixed1 = new global::Gtk.Fixed ();
-		this.fixed1.Name = "fixed1";
-		this.fixed1.HasWindow = false;
-		// Container child fixed1.Gtk.Fixed+FixedChild
-		this.fileChooser = new global::Gtk.FileChooserButton (global::Mono.Unix.Catalog.GetString ("Selecione um arquivo"), ((global::Gtk.FileChooserAction)(0)));
-		this.fileChooser.Name = "fileChooser";
-		this.fixed1.Add (this.fileChooser);
-		global::Gtk.Fixed.FixedChild w1 = ((global::Gtk.Fixed.FixedChild)(this.fixed1 [this.fileChooser]));
-		w1.X = 11;
-		w1.Y = 17;
-		this.Add (this.fixed1);
-		if ((this.Child != null)) {
-			this.Child.ShowAll ();
-		}
-		this.DefaultWidth = 741;
-		this.DefaultHeight = 567;
-		this.Show ();
-		this.DeleteEvent += new global::Gtk.DeleteEventHandler (this.OnDeleteEvent);
-	}
-}
