--- XRITLibraryTest/gtk-gui/gui.stetic
@@ -1,32 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<stetic-interface>
-  <configuration>
-    <images-root-path>..</images-root-path>
-    <target-gtk-version>2.12</target-gtk-version>
-  </configuration>
-  <import>
-    <widget-library name="glade-sharp, Version=2.12.0.0, Culture=neutral, PublicKeyToken=35e10195dab3c99f" />
-    <widget-library name="../bin/Debug/XRITLibraryTest.exe" internal="true" />
-  </import>
-  <widget class="Gtk.Window" id="MainWindow" design-size="741 567">
-    <property name="MemberName" />
-    <property name="Title" translatable="yes">MainWindow</property>
-    <property name="WindowPosition">CenterOnParent</property>
-    <signal name="DeleteEvent" handler="OnDeleteEvent" />
-    <child>
-      <widget class="Gtk.Fixed" id="fixed1">
-        <property name="MemberName" />
-        <property name="HasWindow">False</property>
-        <child>
-          <widget class="Gtk.FileChooserButton" id="fileChooser">
-            <property name="MemberName" />
-          </widget>
-          <packing>
-            <property name="X">11</property>
-            <property name="Y">17</property>
-          </packing>
-        </child>
-      </widget>
-    </child>
-  </widget>
-</stetic-interface>
