--- XRITLibraryTest/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="DotSpatial.Data" version="1.9" targetFramework="net45" />
-  <package id="DotSpatial.Mono" version="1.9" targetFramework="net45" />
-  <package id="DotSpatial.Projections" version="1.9" targetFramework="net45" />
-  <package id="DotSpatial.Serialization" version="1.9" targetFramework="net45" />
-  <package id="DotSpatial.Topology" version="1.9" targetFramework="net45" />
-  <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
-</packages>
