+++ goesdump.userprefs
@@ -1,17 +1,16 @@
-﻿<Properties StartupItem="XRITLibraryTest/XRITLibraryTest.csproj" RefactoringSettings.EnableRefactorings="True">
+﻿<Properties StartupItem="LibraryTest/LibraryTest.csproj" RefactoringSettings.EnableRefactorings="True">
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug Headless|x86" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="goesdump/Models/ProgConfig.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="LibraryTest/Program.cs">
     <Files>
       <File FileName="goesdump/GoesDecoder/FileHandler.cs" Line="1" Column="1" />
       <File FileName="XRIT/Tools/ImageTools.cs" Line="1" Column="1" />
-      <File FileName="XRIT/Tools/Organizer.cs" Line="1" Column="1" />
-      <File FileName="XRITLibraryTest/MainWindow.cs" Line="1" Column="1" />
       <File FileName="goesdump/GoesDecoder/ImageManager.cs" Line="1" Column="1" />
       <File FileName="XRIT/Models/GroupData.cs" Line="1" Column="1" />
       <File FileName="goesdump/ChannelDecoder/DemuxManager.cs" Line="1" Column="1" />
-      <File FileName="goesdump/Main.cs" Line="71" Column="27" />
-      <File FileName="goesdump/HeadlessMain.cs" Line="21" Column="27" />
-      <File FileName="goesdump/Models/ProgConfig.cs" Line="25" Column="40" />
+      <File FileName="goesdump/Main.cs" Line="1" Column="1" />
+      <File FileName="goesdump/HeadlessMain.cs" Line="1" Column="1" />
+      <File FileName="goesdump/Models/ProgConfig.cs" Line="1" Column="1" />
+      <File FileName="LibraryTest/Program.cs" Line="145" Column="33" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
